refactor(transactions): #1 ページコンポーネントを責務ごとにモジュール分割
- `-lib/mock-data.ts`: categories / INITIAL_TRANSACTIONS / events のモックデータを集約
- `-components/transaction-form-fields.tsx`: フォームスキーマ・フック・フィールドUIを抽出
- `-components/transaction-table.tsx`: ユーティリティ関数・テーブルレンダリングを抽出
- `-components/add-transaction-dialog.tsx`: 追加ダイアログを抽出
- `-components/edit-transaction-dialog.tsx`: 編集ダイアログを抽出
- `-components/delete-transaction-dialog.tsx`: 削除確認ダイアログを抽出
- `index.tsx`: 981行 → 82行に削減（state・ハンドラ・レイアウトのみ）
