+++ packages/app/frontend/routes/transactions/-lib/mock-data.ts
@@ -1,0 +1,361 @@
+import type {
+  CategoryColor,
+  CategoryIconType,
+} from '@frontend/components/category/types'
+import type { TransactionType } from '@frontend/lib/types'
+
+// モックデータ：本番ではAPIから取得する
+
+export const categories: {
+  id: string
+  name: string
+  type: TransactionType
+  isSaving: boolean
+  icon: CategoryIconType
+  color: CategoryColor
+}[] = [
+  {
+    id: 'cat-1',
+    name: '食費',
+    type: 'expense',
+    isSaving: false,
+    icon: 'utensils',
+    color: 'red',
+  },
+  {
+    id: 'cat-2',
+    name: '交通費',
+    type: 'expense',
+    isSaving: false,
+    icon: 'bus',
+    color: 'blue',
+  },
+  {
+    id: 'cat-3',
+    name: '外食',
+    type: 'expense',
+    isSaving: false,
+    icon: 'coffee',
+    color: 'orange',
+  },
+  {
+    id: 'cat-4',
+    name: '娯楽・グッズ',
+    type: 'expense',
+    isSaving: false,
+    icon: 'music',
+    color: 'purple',
+  },
+  {
+    id: 'cat-5',
+    name: '衣服',
+    type: 'expense',
+    isSaving: false,
+    icon: 'shirt',
+    color: 'pink',
+  },
+  {
+    id: 'cat-6',
+    name: '日用品',
+    type: 'expense',
+    isSaving: false,
+    icon: 'shopping_cart',
+    color: 'teal',
+  },
+  {
+    id: 'cat-7',
+    name: '美容',
+    type: 'expense',
+    isSaving: false,
+    icon: 'heart_pulse',
+    color: 'pink',
+  },
+  {
+    id: 'cat-8',
+    name: '積立：遠征費',
+    type: 'expense',
+    isSaving: true,
+    icon: 'plane',
+    color: 'blue',
+  },
+  {
+    id: 'cat-9',
+    name: '積立：グッズ',
+    type: 'expense',
+    isSaving: true,
+    icon: 'gift',
+    color: 'purple',
+  },
+  {
+    id: 'cat-11',
+    name: '積立：旅行費',
+    type: 'expense',
+    isSaving: true,
+    icon: 'plane',
+    color: 'teal',
+  },
+  {
+    id: 'cat-12',
+    name: '積立：機材費',
+    type: 'expense',
+    isSaving: true,
+    icon: 'zap',
+    color: 'yellow',
+  },
+  {
+    id: 'cat-13',
+    name: '積立：緊急資金',
+    type: 'expense',
+    isSaving: true,
+    icon: 'piggy_bank',
+    color: 'green',
+  },
+  {
+    id: 'cat-10',
+    name: '給与',
+    type: 'income',
+    isSaving: false,
+    icon: 'wallet',
+    color: 'green',
+  },
+]
+
+export const INITIAL_TRANSACTIONS = [
+  {
+    id: 'tx-1',
+    type: 'income' as TransactionType,
+    amount: 282000,
+    categoryId: 'cat-10',
+    categoryName: '給与',
+    transactionDate: '2026-02-03',
+    name: '2月分給与',
+  },
+  {
+    id: 'tx-2',
+    type: 'expense' as TransactionType,
+    amount: 500,
+    categoryId: 'cat-2',
+    categoryName: '交通費',
+    transactionDate: '2026-02-03',
+    name: '定期外乗車',
+  },
+  {
+    id: 'tx-3',
+    type: 'expense' as TransactionType,
+    amount: 1850,
+    categoryId: 'cat-1',
+    categoryName: '食費',
+    transactionDate: '2026-02-05',
+    name: 'スーパー',
+  },
+  {
+    id: 'tx-4',
+    type: 'expense' as TransactionType,
+    amount: 1200,
+    categoryId: 'cat-3',
+    categoryName: '外食',
+    transactionDate: '2026-02-07',
+    name: 'ランチ',
+  },
+  {
+    id: 'tx-5',
+    type: 'expense' as TransactionType,
+    amount: 30000,
+    categoryId: 'cat-8',
+    categoryName: '積立：遠征費',
+    transactionDate: '2026-02-08',
+    name: '2月分積立',
+  },
+  {
+    id: 'tx-6',
+    type: 'expense' as TransactionType,
+    amount: 6500,
+    categoryId: 'cat-7',
+    categoryName: '美容',
+    transactionDate: '2026-02-10',
+    name: 'カット・カラー',
+  },
+  {
+    id: 'tx-7',
+    type: 'expense' as TransactionType,
+    amount: 2200,
+    categoryId: 'cat-1',
+    categoryName: '食費',
+    transactionDate: '2026-02-12',
+    name: 'スーパー',
+  },
+  {
+    id: 'tx-8',
+    type: 'expense' as TransactionType,
+    amount: 150000,
+    categoryId: 'cat-5',
+    categoryName: '衣服',
+    transactionDate: '2026-02-14',
+    name: 'バレンタインコーデ',
+    eventId: 'ev-1',
+    eventName: 'バレンタインイベント',
+  },
+  {
+    id: 'tx-9',
+    type: 'expense' as TransactionType,
+    amount: 4200,
+    categoryId: 'cat-4',
+    categoryName: '娯楽・グッズ',
+    transactionDate: '2026-02-15',
+    name: 'ぬいぐるみ',
+    eventId: 'ev-1',
+    eventName: 'バレンタインイベント',
+  },
+  {
+    id: 'tx-10',
+    type: 'expense' as TransactionType,
+    amount: 1800,
+    categoryId: 'cat-6',
+    categoryName: '日用品',
+    transactionDate: '2026-02-17',
+    name: '薬局',
+  },
+  {
+    id: 'tx-11',
+    type: 'expense' as TransactionType,
+    amount: 980,
+    categoryId: 'cat-1',
+    categoryName: '食費',
+    transactionDate: '2026-02-18',
+    name: 'コンビニ',
+  },
+  {
+    id: 'tx-12',
+    type: 'expense' as TransactionType,
+    amount: 2800,
+    categoryId: 'cat-2',
+    categoryName: '交通費',
+    transactionDate: '2026-02-20',
+    name: '春ライブ 新幹線',
+    eventId: 'ev-2',
+    eventName: '春ライブ遠征',
+  },
+  {
+    id: 'tx-13',
+    type: 'expense' as TransactionType,
+    amount: 8500,
+    categoryId: 'cat-4',
+    categoryName: '娯楽・グッズ',
+    transactionDate: '2026-02-20',
+    name: 'ライブグッズ',
+    eventId: 'ev-2',
+    eventName: '春ライブ遠征',
+  },
+  {
+    id: 'tx-14',
+    type: 'expense' as TransactionType,
+    amount: 2400,
+    categoryId: 'cat-3',
+    categoryName: '外食',
+    transactionDate: '2026-02-21',
+    name: '遠征ご飯',
+    eventId: 'ev-2',
+    eventName: '春ライブ遠征',
+  },
+  {
+    id: 'tx-15',
+    type: 'expense' as TransactionType,
+    amount: 1600,
+    categoryId: 'cat-1',
+    categoryName: '食費',
+    transactionDate: '2026-02-24',
+    name: 'スーパー',
+  },
+  {
+    id: 'tx-16',
+    type: 'expense' as TransactionType,
+    amount: 3200,
+    categoryId: 'cat-3',
+    categoryName: '外食',
+    transactionDate: '2026-02-26',
+    name: '友人と夕飯',
+  },
+  {
+    id: 'tx-17',
+    type: 'expense' as TransactionType,
+    amount: 1200,
+    categoryId: 'cat-1',
+    categoryName: '食費',
+    transactionDate: '2026-02-28',
+    name: 'コンビニ',
+  },
+  {
+    id: 'tx-18',
+    type: 'expense' as TransactionType,
+    amount: 5500,
+    categoryId: 'cat-4',
+    categoryName: '娯楽・グッズ',
+    transactionDate: '2026-03-05',
+    name: '新グッズ発売',
+    eventId: 'ev-3',
+    eventName: '春グッズ',
+  },
+  {
+    id: 'tx-19',
+    type: 'expense' as TransactionType,
+    amount: 8000,
+    categoryId: 'cat-2',
+    categoryName: '交通費',
+    transactionDate: '2026-03-20',
+    name: '春遠征 新幹線',
+    eventId: 'ev-2',
+    eventName: '春ライブ遠征',
+  },
+  {
+    id: 'tx-20',
+    type: 'expense' as TransactionType,
+    amount: 12000,
+    categoryId: 'cat-4',
+    categoryName: '娯楽・グッズ',
+    transactionDate: '2026-03-20',
+    name: '春ライブグッズ予定',
+    eventId: 'ev-2',
+    eventName: '春ライブ遠征',
+  },
+  {
+    id: 'tx-21',
+    type: 'expense' as TransactionType,
+    amount: 3000,
+    categoryId: 'cat-3',
+    categoryName: '外食',
+    transactionDate: '2026-03-21',
+    name: '遠征飯予定',
+    eventId: 'ev-2',
+    eventName: '春ライブ遠征',
+  },
+  {
+    id: 'tx-22',
+    type: 'expense' as TransactionType,
+    amount: 15000,
+    categoryId: 'cat-5',
+    categoryName: '衣服',
+    transactionDate: '2026-04-05',
+    name: '春物購入予定',
+  },
+]
+
+export const events = [
+  {
+    id: 'ev-1',
+    name: 'バレンタインイベント',
+    dateRange: { start: '2026-02-14' },
+    totalAmount: 12000,
+  },
+  {
+    id: 'ev-2',
+    name: '春ライブ遠征',
+    dateRange: { start: '2026-02-20', end: '2026-03-21' },
+    totalAmount: 35700,
+  },
+  {
+    id: 'ev-3',
+    name: '春グッズ',
+    dateRange: { start: '2026-03-05' },
+    totalAmount: 5500,
+  },
+]
